requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backoff==1.11.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
discord-components==2.1.2
discord.py==1.7.3
idna==3.3
multidict==5.2.0
pip==21.3.1
pycparser==2.21
PyNaCl==1.4.0
requests==2.26.0
setuptools==57.4.0
six==1.16.0
spotify==0.10.2
typing-extensions==3.10.0.2
yarl==1.7.2
youtube-dl==2021.6.6